Created basic template for webpage
Modified app.js to include a basic header and body for the homepage. Modified index.html to direct to root. Created app.css to add formatting to webpage template> Added scripts to package.json to start the web application and for potential testing. Imported React into index.js to render the application